Refactor Streamlit app into multi-page layout
